extracted AasDataLoader Hook
